requirements.txt
requests==2.23.0
pandas==1.0.3
numpy==1.18.4
wordcloud==1.7.0
matplotlib==3.2.1
jieba==0.42.1
snownlp==0.12.3
Pillow==7.1.2
